requirements.txt
streamlit==1.27.2
Pillow==10.0.1
torch==2.1.1
torchvision==0.16.1
pandas==2.1.1
